refactor(ao): extract agents doctor adapter
