+++ cli/cmd/ao/agents.go
@@ -7,12 +7,11 @@ import (
 	"fmt"
 	"os"
 	"path/filepath"
-	"sort"
-	"strings"
 
 	"github.com/spf13/cobra"
 
 	"github.com/boshu2/agentops/cli/internal/adapters/agentslint"
+	"github.com/boshu2/agentops/cli/internal/adapters/agentsurface"
 )
 
 var agentsCmd = &cobra.Command{
@@ -74,11 +73,7 @@ func init() {
 type AgentsLintError = agentslint.Error
 
 // AgentsInventory is the shape returned by `ao agents inspect --json`.
-type AgentsInventory struct {
-	Contract  string   `json:"contract"`
-	Allowlist []string `json:"allowlist"`
-	Skills    []string `json:"skills"`
-}
+type AgentsInventory = agentsurface.Inventory
 
 func runAgentsInspect(cmd *cobra.Command, args []string) error {
 	repoRoot, rootErr := resolveAgentsRepoRoot()
@@ -200,62 +195,10 @@ func shouldResolveAgentsDefaultPath(cmd *cobra.Command, flagName, value, default
 	return !filepath.IsAbs(value) && !cmd.Flags().Changed(flagName) && value == defaultValue
 }
 
-// parseAgentsAllowlist extracts the allowlist between the BEGIN/END
-// markers in the contract doc. Inline `# comment` and blank lines are
-// stripped. The result is sorted and de-duplicated.
 func parseAgentsAllowlist(content string) []string {
-	const beginMarker = "<!-- BEGIN agents-write-surfaces-allowlist -->"
-	const endMarker = "<!-- END agents-write-surfaces-allowlist -->"
-
-	seen := make(map[string]bool)
-	out := []string{}
-	inside := false
-	for _, line := range strings.Split(content, "\n") {
-		if strings.Contains(line, beginMarker) {
-			inside = true
-			continue
-		}
-		if strings.Contains(line, endMarker) {
-			inside = false
-			continue
-		}
-		if !inside {
-			continue
-		}
-		if i := strings.Index(line, "#"); i >= 0 {
-			line = line[:i]
-		}
-		line = strings.TrimSpace(line)
-		if line == "" {
-			continue
-		}
-		if seen[line] {
-			continue
-		}
-		seen[line] = true
-		out = append(out, line)
-	}
-	sort.Strings(out)
-	return out
+	return agentsurface.ParseAllowlist(content)
 }
 
-// discoverActiveSkills returns the names of skills/<name>/ entries that
-// have a SKILL.md file. Result is sorted.
 func discoverActiveSkills(skillsDir string) []string {
-	entries, err := os.ReadDir(skillsDir)
-	if err != nil {
-		return []string{}
-	}
-	out := []string{}
-	for _, e := range entries {
-		if !e.IsDir() {
-			continue
-		}
-		if _, err := os.Stat(filepath.Join(skillsDir, e.Name(), "SKILL.md")); err != nil {
-			continue
-		}
-		out = append(out, e.Name())
-	}
-	sort.Strings(out)
-	return out
+	return agentsurface.DiscoverActiveSkills(skillsDir)
 }
