+++ registry.json
@@ -1,6 +1,6 @@
 {
   "schema_version": 2,
-  "generated_at": "2026-06-07T09:59:15Z",
+  "generated_at": "2026-06-07T10:20:50Z",
   "summary": {
     "skills": 173,
     "hooks": 0,
